+++ examples/apps/ui-data-grid/vite.config.js
@@ -1,6 +1,7 @@
 import path from "node:path"
 import { fileURLToPath } from "node:url"
 
+import { hmr } from "@inglorious/vite-plugin-hmr"
 import { minifyTemplateLiterals } from "rollup-plugin-minify-template-literals"
 import { defineConfig } from "vite"
 
@@ -9,6 +10,8 @@ const __dirname = path.dirname(__filename)
 
 // https://vite.dev/config/
 export default defineConfig({
+  plugins: [hmr()],
+
   build: {
     rollupOptions: {
       plugins: [minifyTemplateLiterals()],