+++ examples/apps/web-charts/vite.config.js
@@ -1,11 +1,26 @@
-import { resolve } from "path"
+import path from "node:path"
+import { fileURLToPath } from "node:url"
+
+import { hmr } from "@inglorious/vite-plugin-hmr"
+import { minifyTemplateLiterals } from "rollup-plugin-minify-template-literals"
 import { defineConfig } from "vite"
+
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 
 // https://vite.dev/config/
 export default defineConfig({
+  plugins: [hmr()],
+
+  build: {
+    rollupOptions: {
+      plugins: [minifyTemplateLiterals()],
+    },
+  },
+
   resolve: {
     alias: {
-      "@": resolve(__dirname, "src"),
+      "@": path.resolve(__dirname, "src"),
     },
   },
 })