+++ examples/apps/web-todomvc-jsx/vite.config.js
@@ -1,6 +1,7 @@
 import path from "node:path"
 import { fileURLToPath } from "node:url"
 
+import { hmr } from "@inglorious/vite-plugin-hmr"
 import { jsx } from "@inglorious/vite-plugin-jsx"
 import { minifyTemplateLiterals } from "rollup-plugin-minify-template-literals"
 import { defineConfig } from "vite"
@@ -10,7 +11,7 @@ const __dirname = path.dirname(__filename)
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [jsx()],
+  plugins: [hmr(), jsx()],
 
   build: {
     rollupOptions: {