+++ benchmarks/dashboard/react-inglorious-memo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @benchmarks/dashboard-react-inglorious-memo
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@11.0.1
+
 ## 0.1.3
 
 ### Patch Changes