+++ examples/apps/react-todomvc-rtk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # react-todomvc-rtk
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@11.0.1
+
 ## 0.1.4
 
 ### Patch Changes