+++ examples/apps/ui-data-grid/CHANGELOG.md
@@ -1,5 +1,15 @@
 # lit-todomvc
 
+## 0.1.58
+
+### Patch Changes
+
+- Add HMR Vite plugin
+- Updated dependencies
+  - @inglorious/store@11.0.1
+  - @inglorious/web@6.4.1
+  - @inglorious/ui@7.1.1
+
 ## 0.1.57
 
 ### Patch Changes