+++ examples/apps/web-charts-ssx/CHANGELOG.md
@@ -1,5 +1,15 @@
 # web-charts-ssx
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@11.0.1
+  - @inglorious/web@6.4.1
+  - @inglorious/ssx@2.1.5
+  - @inglorious/charts@10.0.1
+
 ## 0.1.1
 
 ### Patch Changes