+++ examples/apps/web-charts/CHANGELOG.md
@@ -1,5 +1,15 @@
 # web-charts
 
+## 0.1.18
+
+### Patch Changes
+
+- Add HMR Vite plugin
+- Updated dependencies
+  - @inglorious/store@11.0.1
+  - @inglorious/web@6.4.1
+  - @inglorious/charts@10.0.1
+
 ## 0.1.17
 
 ### Patch Changes