+++ examples/apps/web-components/CHANGELOG.md
@@ -1,5 +1,14 @@
 # web-components
 
+## 0.0.15
+
+### Patch Changes
+
+- Add HMR Vite plugin
+- Updated dependencies
+  - @inglorious/store@11.0.1
+  - @inglorious/web@6.4.1
+
 ## 0.0.14
 
 ### Patch Changes