+++ examples/apps/web-finance-dashboard/CHANGELOG.md
@@ -1,5 +1,16 @@
 # web-finance-dashboard
 
+## 0.1.10
+
+### Patch Changes
+
+- Add HMR Vite plugin
+- Updated dependencies
+  - @inglorious/store@11.0.1
+  - @inglorious/web@6.4.1
+  - @inglorious/ui@7.1.1
+  - @inglorious/charts@10.0.1
+
 ## 0.1.9
 
 ### Patch Changes